fix: address review feedback — MAIN_ROLE to constants.py, RoleUsage to ttypes.py
